This commit was manufactured by cvs2svn to create tag 'wxPy_2_6_2_1'.

+++ src/motif/toplevel.cpp
@@ -1,5 +1,5 @@
 /////////////////////////////////////////////////////////////////////////////
-// Name:        src/motif/toplevel.cpp
+// Name:        motif/toplevel.cpp
 // Purpose:     wxTopLevelWindow Motif implementation
 // Author:      Mattia Barbon
 // Modified by:
@@ -16,6 +16,10 @@
 // ----------------------------------------------------------------------------
 // headers
 // ----------------------------------------------------------------------------
+
+#if defined(__GNUG__) && !defined(NO_GCC_PRAGMA)
+    #pragma implementation "toplevel.h"
+#endif
 
 // For compilers that support precompilation, includes "wx.h".
 #include "wx/wxprec.h"
@@ -65,7 +69,14 @@ static void wxTLWEventHandler( Widget wid,
 
 void wxTopLevelWindowMotif::PreDestroy()
 {
-    wxModelessWindows.DeleteObject(this);
+#ifdef __VMS
+#pragma message disable codcauunr
+#endif
+   if ( (GetWindowStyleFlag() & wxDIALOG_MODAL) != wxDIALOG_MODAL )
+        wxModelessWindows.DeleteObject(this);
+#ifdef __VMS
+#pragma message enable codcauunr
+#endif
 
     m_icons.m_icons.Empty();
 
@@ -110,7 +121,7 @@ bool wxTopLevelWindowMotif::Create( wxWindow *parent, wxWindowID id,
 
     m_windowId = ( id > -1 ) ? id : NewControlId();
 
-    bool retval = XmDoCreateTLW( parent, id, title, pos, size, style, name );
+    bool retval = DoCreate( parent, id, title, pos, size, style, name );
 
     if( !retval ) return false;
 
@@ -263,8 +274,8 @@ WXWidget wxTopLevelWindowMotif::GetShellWidget() const
     return (WXWidget) GetShell( this );
 }
 
-bool wxTopLevelWindowMotif::ShowFullScreen( bool WXUNUSED(show),
-                                            long WXUNUSED(style) )
+bool wxTopLevelWindowMotif::ShowFullScreen( bool show,
+                                            long style )
 {
     // TODO, see wxGTK
     return false;
@@ -311,7 +322,7 @@ bool wxTopLevelWindowMotif::IsIconized() const
                    XmNiconic, &iconic,
                    NULL );
 
-    return (iconic == True);
+    return iconic;
 }
 
 void wxTopLevelWindowMotif::Maximize( bool maximize )
